Initial proto objects for metric server hosted in a runtime app instance
An http server may be deployed inside an eve app instance which hosts
http endpoints returning nested application instance metrics.
These objects are returned in the body of the http GET so that
eve knows which application instance ids are running nested
inside the runtime and the metrics for each of those respective
nested app instances.

Signed-off-by: Andrew Durbin <andrewd@zededa.com>